+++ packages/api/src/Domain/Client/User/UserApiServiceInterface.ts
@@ -1,5 +1,5 @@
 import { UserRequestType, Uuid } from '@standardnotes/common'
-import { RootKeyParamsInterface } from '@standardnotes/models'
+import { type RootKeyParamsInterface } from '@standardnotes/models'
 
 import { UserDeletionResponse } from '../../Response/User/UserDeletionResponse'
 import { UserRegistrationResponse } from '../../Response/User/UserRegistrationResponse'