+++ packages/mobile/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [3.46.21](https://github.com/standardnotes/app/compare/@standardnotes/mobile@3.46.20...@standardnotes/mobile@3.46.21) (2022-11-18)
+
+**Note:** Version bump only for package @standardnotes/mobile
+
+## [3.46.20](https://github.com/standardnotes/app/compare/@standardnotes/mobile@3.46.19...@standardnotes/mobile@3.46.20) (2022-11-17)
+
+**Note:** Version bump only for package @standardnotes/mobile
+
 ## [3.46.19](https://github.com/standardnotes/app/compare/@standardnotes/mobile@3.46.18...@standardnotes/mobile@3.46.19) (2022-11-17)
 
 **Note:** Version bump only for package @standardnotes/mobile